Transfer default fixups
Default to "NONE" for all IN parameters; this more closely matches
how we did transfer in the old system.  However - we need to make
an exception for strings (utf8).  For some reason those were
FULL, even though that has no meaning.

Also default to NONE for void to match the old system.


svn path=/branches/annotation/; revision=1011