Bug 575613 - Enum stripping with common prefix, also use "_" consistently
Some enums have members which have a common prefix which doesn't
match that of the enum name, but it also longer than the global
namespace prefix.  Instead, try stripping the common prefix first,
and only if that fails fall back to the global strip.

Also, for glib-registered enums we were using the nick, which
typically has "-" as a separator.  Replace that with "_" for
consistency between unregistered enums and registered.

utils.py:strip_common_prefix is now unused, delete.