Use all libraries specified on the command line
Earlier we were only using the first library specified when
compiling an introspection binary.  This breaks builds on Solaris
which doesn't have implicit linking and we need to be able
to specify multiple libraries.

The tension here is primarily related to OS X' distinction between
modules and shared libraries; if we build the custom libraries
in gir-repository as shared libraries, then we can't dlopen them.
If we build them as modules, then the introspection binary compiler
can't link to them.

If we want to fix the OS X build, probably the way to go would be
a special --use-first-library-only option that gets conditionalized
on OS X + gir-repository.

But even better is to finish killing gir-repository entirely.