Fix glib:error-quark scanning for unregistered enum types
The enumeration type for an error might not be introspected
and registered with GObject (this is, in particular, true for
the errors in GLib itself). So, as a fallback search, through
all enum types in the module.

Add a special case for the aberrant GIOErrorEnum and a warning
if resolution completely fails.

https://bugzilla.gnome.org/show_bug.cgi?id=602512