Makefile.introspection: use $^, simplifying non-srcdir builds
In the .gir-building rule, use "$^" to refer to the source files,
since that automatically looks in both $(srcdir) and $(builddir). This
is particularly important since certain generated files will be in
$(builddir) when building from git, but in $(srcdir) when building
from tarballs

If you were previously prefixing $(srcdir) to the Foo_gir_FILES
members by hand, you should stop now.

(Also, removed the dependencies on $(INTROSPECTION_SCANNER) and
$(INTROSPECTION_COMPILER) for the .gir/.typelib rules, since the
scanner one was broken anyway, and we don't have that kind of
dependency for other rules (eg, making .o files depend on
/usr/bin/gcc).)

https://bugzilla.gnome.org/show_bug.cgi?id=616425