Go back to previously handling of static methods on non-GObjects
Previous to the scanner rewrite, static methods were only handled
for GObject subclasses. Go back to that for full compatibility
with existing bindings and code. See bug 572408 for the topic
of changing the API.

The code for writing out static methods for records and boxed is
left as future proofing, and the same handling is added for
interfaces.

https://bugzilla.gnome.org/show_bug.cgi?id=628967