Handle casing better for constants
Instead of handling constants by lower-casing them, stripping the
lower-case prefix and upper-casing them again, leave them in the
original case and check against upper-cased versions of
namespace.symbol_prefixes.

Wwe detect what version to test against by looking at the first character
of the identifier, so we assume that --symbol-prefix options are always
in lowercase. If that needs to be relaxed, then we'll have to check all
symbols against both sets of prefixes.

Add tests for constants to Regress.h and fix tests/warn/unresolved-type.h
for a warning message that improved with this change.

https://bugzilla.gnome.org/show_bug.cgi?id=629007