[scanner] Only store types in cache store
Only store types that can be referenced in the cache store
this reduces the size of the serialized Gtk-3.0.gir in the store
from 5.7M to 366k on my system.

It also reduces the time it takes to create a gir in gtksourceview
by 35% and the time to run the warnings test by more than 50%

This also disables the cache for passthrough mode since it needs
access to the whole serialized tree.