dumper: Differentiate between "external" and "internal" linking
The previous commit aea515709 broke scanning for libraries not
in the current directory, such as scanning Gio from gobject-introspection.

In this case, it's wrong to add -L., and to move the other -L behind
the library.  Instead, we should just do a "normal" link using
pkg-config --libs with few games.

https://bugzilla.gnome.org/show_bug.cgi?id=632701