scanner: support typedefs for container types
Inside an alias definition, we only need the name of the alias target
type.  Add a method to GIRWriter to write out a type reference rather
than full type definition and use it when writing out an alias.

Determine introspectableness of aliases in IntrospectablePass so functions
using aliases can be marked not-introspectable if the alias itself is not.

https://bugzilla.gnome.org/show_bug.cgi?id=629682