scanner: use c-symbol-prefix for method pairing where possible
If we have a C symbol prefix for the target type, then use that to
identify methods that should be paired with it.  The C symbol prefix
is derived from the name of the _get_type() function, so it's likely
that this is also the prefix given to methods for the type.

https://bugzilla.gnome.org/show_bug.cgi?id=634153