cairo-1.0.gir: Un-foreign RectangleInt
cairo.RectangleInt is a very tricky case; it's (as of now)
boxed in cairo-gobject.  We're trying to fix a few bugs here.  First,
we want to make (out caller-allocates) work.  In order to do that,
we need to explicitly specify the fields so that the typelib
has the right size.

Also, add the boxed type now so that we pick up the right thing
from GObject signal parameters.

See history in bug 623222 and bug 636393.

https://bugzilla.gnome.org/show_bug.cgi?id=636617