Add a second enum to GIMarshallingTests to support pygi tests for non-gtype enums.
There was a bug in pygi in which non-gtype enums shared values between themselves.
That is, if there was an attribute Enum.THE_VALUE, it would be available on as OtherEnum.THE_VALUE as well.
To make a test case for this bug, we need a second non-gtype enum in GIMarshallingTests to compare with.