scanner: Also add an RPATH for library paths specified
Introspection scanning is a bit special in that we both compile *and*
run a temporary binary.  In some cases like jhbuild we will pick up
the runtime location of libraries via LD_LIBRARY_PATH.  But this falls
over when we've been given a private subdirectory, since it won't be
in that path.

Let's just "do the right thing" here by injecting an rpath for the
directories as well.