Fix regress_test_closure_variant return value
We need to g_variant_ref() the return type: the closure delivers it wrapped
in a GValue, and the subsequent g_value_unset() unrefs it to zero otherwise.

This was uncovered by the now working TestClosures.test_variant() test case in
PyGObject.

Also allow passing NULL as argument, to be able to test these code paths as
well.