scanner: allow for functions that look like constructors but aren't
If the scanner found a function with _new in its name, but not
prefixed by the name of a known type, it would assume that the
function was a constructor, and then complain that it couldn't figure
out what it was a constructor for, and mark it introspectable=0.

Instead, just assume that the function is not actually a constructor
in that case (unless it's explicitly tagged as such).

https://bugzilla.gnome.org/show_bug.cgi?id=676815