sourcescannermain.py: Add --filelist option
Add a function to put the files to pass to g-ir-scanner in a filelist
file, with one source/header file on a single line.  This is done to
work around the limitation of Windows cmd.exe (and MSYS) where the command
line cannot exceed 8192 characters.

When this option is used, g-ir-scanner will only look for the header/source
files that are specified in the filelist file, which is not too hard to
generate.

https://bugzilla.gnome.org/show_bug.cgi?id=681820