Add Visual Studio projects
These are the Visual C++ 2008/2010 project files that can be used to build
gobject-introspection.  The main girepository dll (along with cmph), tools
(in C), certain test/sample programs dlls are built with this project file
set.

Due to a bug in the Visual C++ 2008 linker, the g-ir-compiler and
g-ir-generator tools have the libgirepository-internals sources built
along side with the sources of these respective tool programs.

In order to ease future maintenance, the cmph, girepository,
g-ir-compiler and property sheets (for the "install" part) are done as
templates where header/source listings are filled in during "make dist",
so that the headers/sources listings are kept up to date between releases.

The GLib and girepository .gir and .typelib files are also generated and
"compiled" as part of the build process, using gengir.bat that
was committed earlier.

https://bugzilla.gnome.org/show_bug.cgi?id=681820