build: Ensure cairo-gobject always ends up in cairo.gir
We were conditionalizing on the presence of cairo in the buildroot,
but that's super lame since it makes a huge mess of the build
dependency graph.

Let's just always stick cairo-gobject in there.  I doubt anyone cares
anymore about building g-i on really old systems without it.

If they do...well, we can revisit this again.