.gitignore: Ignore most .page files
While testing g-ir-doctool, it's common to generate a directory
full of a large number of small .page files inside the source
directory, which slows down git a lot while searching for untracked
files. Ignore the .page files for now, so git doesn't have to look
in its database for every one.