Visual C++ 2010 projects: Prepare for VS 2012 support
Add the PlatformToolset tag to the project configs so that we can use
a simple script later to the autotools files to copy the projects and
change the value of that tag (v100 -> v110, and other simple changes)
in order that we can quickly provide and maintain support for Visual
Studio 2012 with minimal effort.

Note that at the moment G-I does not yet support the API/SDK requirements
for Windows 8 Modern UI (formerly known as Metro), but this paves the very
initial step.