doctool: Rename templates to exclude mallard/the language
As templates are in their own directory and segregated into language
already, this is sort of repeating the issue. At the same time, always
explicitly use relative ("./") or absolute ("/") lookups for templates.

We want to eventually have base templates to share between languages, so
to do so without namespace clashes makes sense.