giscanner: fix string formatting tuples
During review Colin Walters mentioned:

"""
One thing I should probably explain is my habit of *always* passing a
tuple for %.  The reason is pretty simple...a long time ago I had an
error handler that did:  print "unexpected data: %s" % foo.  And I'd
changed the type of "foo" from a string to a tuple at some point.
But this obviously caused "print" to fail because it was expecting
multiple format strings.

Basically always using tuples is safer in case you change the type
of a variable to a tuple.
"""

This fixes violations of this rule introduced up until now.

https://bugzilla.gnome.org/show_bug.cgi?id=697614