scanner: internals cleanup: Key more things off Namespace
The .gir format has a weird legacy where stuff like the includes are
outside of the <namespace>.  But conceptually they're tied together,
so let's start reflecting this in the code.

This way we can just pass around and look at a Namespace object
instead of a 4-tuple of (namespace, includes, c_includes, pkg_config).

https://bugzilla.gnome.org/show_bug.cgi?id=694593