Visual Studio builds: Process cairo-1.0.gir.in
Use a simple Python script that uses Python REGEX abilities to create
cairo-1.0.gir from cairo-1.0.gir.in with the correct values for Cairo, for
use in the Visual Studio builds of g-i.  It can be later extended if
needed.

Also dist this, and call this script from gengir.bat so that cairo-1.0.gir
and cairo-1.0.typelib can be built and "installed"

https://bugzilla.gnome.org/show_bug.cgi?id=692165