giscanner: simplify the MainTransformer().transform() method
The ast.Namespace.names instance attribute is an OrderedDict
which itself is a specialised dict() subclass. There are any
number of ways to test if a dict() is empty or not. Creating
a copy of it's keys by iterating over them just to count the
number of items in the copy is not the most elegant way though.

https://bugzilla.gnome.org/show_bug.cgi?id=699533