scanner: Consistently use realpath() on input filenames
The scanner has some logic to compare the filenames specified
on the command line against files it will parse.  For the latter,
it uses g_realpath().  With this patch, we also use g_realpath()
on the command line arguments.

This fixes g-i when used inside jhbuild in a gnome-ostree VM,
which has a symbolic link /home -> /sysroot/home.  This caused
a realpath mismatch, and then we'd ignore the input source files.

It'd be best to get out of the realpath business entirely...but
a patch to do that seems more likely to break.

https://bugzilla.gnome.org/show_bug.cgi?id=704864