giscanner: Handle the case when there's a space between -I cpp flag and the path
In this case we were considering -I and the path as two different
command line options. This was not a problem in the past (or for other
cpp flags) because they are passed directly to gcc that accepts it. Now
that we are ensuring that the include paths are always real paths, we
need to handle this case to identify include paths in the command line.

https://bugzilla.gnome.org/show_bug.cgi?id=712211