Makefile.introspection: pass *FLAGS to scanner
automake provides a few mechanisms for setting the various user-defined
variables CFLAGS, CPPFLAGS, LDFLAGS that g-ir-scanner passes through to
the C compiler.

 - can be set in the environment

 - can be given as a commandline option to configure:

     ./configure LDFLAGS=-L/usr/local/lib

 - can be given as a commandline option to make:

     make LDFLAGS=-L/usr/local/lib

Currently, g-ir-scanner only checks for the variables in the
environment, whereas the correct value to use will always be the value
that is available from the Makefile, eg: $(LDFLAGS).

We could pass these variables as new arguments to the scanner, but the
easiest thing to do is promote them to the environment where the scanner
can see them.

https://bugzilla.gnome.org/show_bug.cgi?id=722887