docwriter/gjs: fix constructors
We should only look at introspectable constructors when resolving
the default and zero args. Also, a disguised structure or one
with zero fields can't be allocated directly.

In addition, restore availability of all constructors, because
we need a page to document the default one, and because some classes
have problems with GObject construction.