docwriter: improve xrefs
If a xref points to a node that was not visited, because it
was not introspectable, don't assume it has no parent. Instead,
try to fetch the parent from the node directly.
This is still wrong (the xref points nowhere because no page is
generated), but it looks less wrong.
If OTOH a xref points to a different namespace, generate the
link directly, because yelp-build is not capable of translating
it.