tests: Change names with case-only distinction
RegressIntSet and RegressIntset, aliases of one another, were causing
two documentation pages to be built by the documentation tests that
were named identically except for letter case.

On OSX and Windows, the most common filesystems are case-insensitive,
and those two files can't exist in the same directory on case-
insensitive filesystems. That caused the tests to fail. It also caused
problems when checking out the repository, because the expected
documentation set also contains those files.

This commit fixes the symptom (so that JHbuild will stop erroring out
on this module) but not the problem of generating a documentation set
for two identifiers which differ (perfectly legally) only by case.

https://bugzilla.gnome.org/show_bug.cgi?id=725264