build/gi-introspection-msvc.mak: Fix Build and Installation
Copy the .gir files that are bundled with g-i to build/ before trying to
compile the .gir's that were generated during the build process, so that
the compilation of these .gir's may succeed.  Also, create the
share/gir-1.0 and lib/girepository-1.0 directories, if needed, before
trying to copy the .gir/.typelib files.