giscanner: write nullable and optional attributes
Record our internal 'nullable' and 'optional' attributes into the
written .gir file.  It is now theoretically possible to express the
concept of an out parameter with a nullable type (although presently
there is no way to do this).

Modify our own internal parser (in the scanner) to understand the
newly-written attributes.

Update the expected output of the 'Regress-1.0.gir' test to account for
the new attributes.

Nothing else understands 'nullable' yet, but the girparser in the
typelib compiler already understands 'optional' and records a bit for it
in the typelib.

https://bugzilla.gnome.org/show_bug.cgi?id=660879