giscanner: change some internal field logic
Replace the 'allow_none' field on parameters with two separate fields:
'nullable' and 'optional'.

Currently, we use 'nullable' to mean the same thing as 'allow-none' for
normal (non-out) parameters.  For out parameters, we use the 'optional'
field instead.

Note that the special case for GCancellable and GAsyncReadyCallback is
already guarded by a check for being an in parameter, so we always use
'nullable' here.

On the .gir writer side, we decide which variable to consult when
writing the allow-none attribute depending on the parameter direction.

https://bugzilla.gnome.org/show_bug.cgi?id=660879