giscanner: tweak GCancellable null special case
We treat GCancellable and GAsyncReadyCallback as if they were always
tagged (allow-none) because these parameters universally allow %NULL to
be given as a value when they appear as in parameters on functions.

Unfortunately, the meaning of (allow-none) is different on out
parameters.  There it means that you can give %NULL to ignore the return
value.

Limit this special case to only in parameters.

https://bugzilla.gnome.org/show_bug.cgi?id=660879