giscanner: Make _get_cachedir() Always Work on Windows
On Windows, checking for $(HOME) will work in a MSYS shell but not in
cmd.exe (i.e. Visual Studio command prompt), so we need to check for
HOMEPATH when we are building under a Visual Studio command prompt to get
the users's home directory correctly.  This will enable g-ir-doc-tool to
work on Windows when run from cmd.exe.

https://bugzilla.gnome.org/show_bug.cgi?id=732668