MSVC Builds: Add NMake Makefiles to Build the Tests
This adds NMake Makefiles for one to build the tests for Visual Studio
builds, which is not done with an MSYS shell.  Shared items from the
existing NMake Makefiles for building the introspection files, which are
also used to build the tests, are split out into common NMake Makefiles,
which are then included.

This also means that we do not need to include the test_gir_cmd.txt for
people to look at when building the tests as the NMake Makefiles will do
the job for them, which cover the material there.