build: Export Symbols Using Compiler Directives
Use compiler directives for exporting symbols for the build of
libgirepository and also for the test libraries, like what is now done
in GLib and GTK+ so that maintaining a separate .symbols (and .def files)
would not be needed, in which the correct compiler directive is determined
during configure time.  Drop all the .def files and the .symbols file as
we would not be using them anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=732669