cmph/bdz.c: Work Around MSVC 2012 x64 Compiler Bug
Due to an MSVC 2012 x64 compiler issue, the compiler generates bad code
for bdz.c, so the for loop in assign() continues running until the point
i falls below zero, causing an access violation when we try to do
curr_edge=queue[i]; (line 427 in bdz.c).  Address this issue by breaking
out of the loop at the end of it when i reaches 0 after doing the
necessary processing.

https://bugzilla.gnome.org/show_bug.cgi?id=733595