tests: Add a Header for Symbol Exporting
This adds a header under tests/ that can be included by the various test
libraries/sources so that symbols from them can be decorated by a macro,
which can then be used to export symbols by using a compiler directive,
like the updates to girepository.  This header is also installed as the
test headers refer to them.

https://bugzilla.gnome.org/show_bug.cgi?id=732669