giscanner: Add Optional Options for Codegen
This adds options to scannermain so that we can decorate functions with
macros as needed, so that we can use compiler annotations for symbol export
for example.

Options are also added to include headers before or after the main include
block at the top so that we can include headers as necessary in the
generated sources and/or headers, so that we could for example grab
definitions from those headers as needed, such as to grab definitions of
macros used for symbol export.

The testcodegen.py script has been updated as well to make use of this
functionality, if needed.

https://bugzilla.gnome.org/show_bug.cgi?id=732669