build: use the detected pkg-config
Rather than assuming that pkg-config is spelt pkg-config, use the environment
variable $PKG_CONFIG which is the actual detected pkg-config spelling.  This is
particularly important when cross-compiling, where the pkg-config may be
prefixed with the target and will actually provide a different result from the
build pkg-config which may be available unprefixed.

https://bugzilla.gnome.org/show_bug.cgi?id=746669