MSVC .gir Builds: Allow More Flexibility for Cairo
Allow one to pass in the cairo-gobject DLL filename for generating
cairo-1.0.gir, as the DLL filename for Cairo-GObject may vary between
different setups, if needed.  Display a note telling the user that he/she
needs to ensure that the DLL name used is correct so that programs and
scripts using the cairo-1.0.typelib can run correctly.

Also clean up the generation of cairo-1.0.gir a bit by directly generating
the cairo-1.0.gir in build\ only during the build.