MSVC Builds: Update to Fix Build of Scripts
Update the custom build steps that is used to generate the scripts in
tools/, so that the Visual Studio builds can be done successfully.

Also assume that one will be using the same Python release series for x86
and x64 builds, while we still leave room for one to customize these if
necesssary.