MSVC Builds: "Add" Visual Studio 2015 Projects
Update the common autotools module that was used to create the
Visual Studio 2012/2013 projects so that we can create the 2015
projects as well.  The 2015 projects, like the 2012 and 2013 ones
are largely similar to the 2010 ones in format, so we just need to
copy the 2010 projects and replace the items in there as needed.