giscanner: fix alias resolution
Atk has a "typedef GSList AtkAttributeSet", which causes
an ast.Alias() targeting a ast.List(). The latter has a
target_fundamental of '<list>', which cannot be looked
up in ast.type_names (because it's not a real fundamental
type), and that breaks the build.

It turns out that we don't need to find the fundamental
type and then the fundamental type node - the target
field of the ast.Alias already points to the desired
type node.

https://bugzilla.gnome.org/show_bug.cgi?id=755681