build: Add an autotools module for introspection builds on MSVC
This adds an autotools module that can be included in projects to be
used for generating the command lines for g-ir-scanner and
g-ir-compiler for introspection builds on Visual Studio builds.

This is currently not used directly in the MSVC builds of g-i, but
is added here as other modules are directly to get this file from
here (i.e. $(srcroot)/build) for inclusion in their Makefile.am's
to support introspection builds for modules that support Visual
Studio builds.