giscanner/ccompiler.py: Make Windows builds work with Python 3.x
Windows builds check the library that was passed into the introspection
scanner to deduce the correct DLL that is to be used by the .gir/.typelib
file, but this was not updated for Python 2.x/3.x compatibility, as the
outputs of a subprocess must be decoded.  Fix this for both MSVC and
MinGW-based builds.

https://bugzilla.gnome.org/show_bug.cgi?id=757126