build: Move MSVC-related build items to build/win32
Move the NMake Makefiles modules and the Python tool scripts to build/win32,
so that we may be consistent across the board.  Also, update the build files
and the README.txt's to make it clear to people that Python 3.3+ is now
supported for building the introspection files, in addition to Python 2.7.x.

Make the NMake Makefiles more flexible in powershell environments so that
the variables can be properly passed in from the command line.