devdocs: Add deprecated CSS class where needed
This adds a def in _doc.tmpl that expands to "deprecated" if the node is
deprecated, and to nothing otherwise. We use it to give the deprecated CSS
class to particular elements. The presence of the class on a symbol's main
entry will inform DevDocs that the symbol is deprecated, though it won't
be formatted in a particular way. The presence of the class on an index
entry will cause the index entry to be struck out, so that it's visually
obvious in the index when a symbol is deprecated.