devdocs: Mark links to other GIRs as such
DevDocs will try to normalize links relative to the current documentation
if it can, and strips ../ off the front, so e.g. a link on the
gio20/gio.file page to ../gobject20/gobject.object will end up resolving
to http://<host>/gio20/gobject20/gobject.object, which doesn't exist.

We clearly mark links to other GIRs with a gir:// URL scheme that DevDocs
won't process automatically, and then fill in the links ourselves in the
DevDocs scraper.