MSVC builds: Make introspection builds a bit easier
GLib recently had a script that would generate the pkg-config files for
its Visual Studio builds, which would be sufficient for the need of
building introspection files, so update the NMake Makefile modules to
look for them automatically, without the need to manually set
PKG_CONFIG_PATH in the console, provided that the build was done in the
same directory tree[1] and was not moved.  Note that this would precede
any PKG_CONFIG_PATH's that were already set in the console.

[1]: This would mean:
$(tree_root)
    |
    |--$(GLib_src_tree)
    |--$(G-I_src_tree)
    |...